Updating unit tests to be more stable + minor cleanup
