+++ .github/workflows/gradle.yml
@@ -1,0 +1,26 @@
+# This workflow will build a Java project with Gradle
+# For more information see: https://help.github.com/actions/language-and-framework-guides/building-and-testing-java-with-gradle
+
+name: Android CI
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - name: Java Setup
+      uses: actions/setup-java@v1
+      with:
+        java-version: 8
+    - name: Grant execute permission for gradlew
+      run: chmod +x gradlew
+    - name: Unit Test
+      run: ./gradlew testAndroidDebugUnitTest
+    - name: Lint
+      run: ./gradlew lintAndroidDebug
